requirements.txt
asgiref==3.6.0
certifi @ file:///Users/ec2-user/ci_py311/certifi_1678321191472/work/certifi
Django==4.2
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
